requirements.txt
#NLP2 Requirements

# Streamlit - Interfaz web
streamlit>=1.28.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-anthropic>=0.1.0
langchain-pinecone>=0.1.0
langchain-huggingface>=0.0.1

# Vector database
pinecone-client>=3.0.0

# Embeddings
sentence-transformers>=2.2.0

# PDF processing
PyPDF2>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# Groq API
groq>=0.4.0